Extracts audit code to concern for Dog audits
