+++ app/controllers/dogs_controller.rb
@@ -95,6 +95,8 @@ class DogsController < ApplicationController
   end
 
   def update
+    @dog.updated_by_admin_user= current_user
+
     if @dog.update_attributes(dog_params)
       flash[:success] = 'Dog updated.'
       redirect_to @dog
